backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openai==0.28
python-dotenv==1.0.0